requirements.txt
Flask==2.2.2
pytest==7.2.1
PyJWT==2.6.0
faiss-cpu==1.7.3
sentence-transformers==2.2.2